Set current line highlight for line numbers
This sets the same highlight background + boldface in the number line
that is already used for the rest of the current line. Previously, the
current line background only covered the line itself and not the line
numbers column and the colour was being set manually to 15 rather than
just using 'bold'.